Use double for teacher hours and improve audit logging
Changed duration and hour-related fields from long to double in ClassSchedule, AvailableTeacherDTO, and TeacherDashboardDTO to support fractional hours. Updated TeacherDashboardService logic to use double for hour calculations. Added audit logging for academic period closure requests and approvals in AcademicPeriodService.

+++ backend-preu/src/main/java/com/backend/backendpreu/users/dto/AvailableTeacherDTO.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class AvailableTeacherDTO {
     private Long id;
     private String fullName;
-    private long currentHours;
+    private Double currentHours;
     private String status;
 }
