+++ backend-preu/src/main/java/com/backend/backendpreu/users/service/TeacherDashboardService.java
@@ -31,15 +31,11 @@ public class TeacherDashboardService {
 
     @Transactional(readOnly = true)
     public List<TeacherDashboardDTO> getTeachersDashboard() {
-        // 1. Traer todos los usuarios que sean PROFESSOR
-        List<User> teachers = userRepository.findByRole(Role.PROFESSOR); // Asegúrate de tener este método en UserRepository
-
+        List<User> teachers = userRepository.findByRole(Role.PROFESSOR);
         return teachers.stream().map(this::buildTeacherStats).collect(Collectors.toList());
     }
 
     private TeacherDashboardDTO buildTeacherStats(User teacher) {
-        // 2. Buscar cursos ACTIVOS donde este usuario sea MAIN_PROFESSOR
-        // (Nota: Debes crear este método en tu ParticipationRepository si no existe, o usar uno genérico)
         List<CourseParticipation> activeParticipations = participationRepository
                 .findAllByUserIdAndRolesInAndStatus(
                         teacher.getId(),
@@ -47,19 +43,21 @@ public class TeacherDashboardService {
                         ParticipationStatus.ACTIVE
                 );
 
-        long totalHours = 0;
+        // CAMBIO 1: Usamos double para decimales (1.5 horas)
+        double totalHours = 0.0;
         List<TeacherDashboardDTO.AssignedCourseInfo> courseInfos = new ArrayList<>();
 
         for (CourseParticipation p : activeParticipations) {
             AcademicPeriod period = p.getAcademicPeriod();
             ClassSchedule schedule = period.getSchedule();
 
-            // Calculamos info del horario
             String scheduleStr = "Sin Horario";
-            long duration = 0;
+            // CAMBIO 2: Variable local double
+            double duration = 0.0;
 
             if (schedule != null) {
                 scheduleStr = schedule.getDayOfWeek() + " " + schedule.getStartTime() + " - " + schedule.getEndTime();
+                // CAMBIO 3: Se asume que getDurationInHours devuelve double
                 duration = schedule.getDurationInHours();
             }
 
@@ -72,7 +70,6 @@ public class TeacherDashboardService {
                     .build());
         }
 
-        // Determinar estado de carga (Lógica simple)
         String status = "LIBRE";
         if (totalHours > 0) status = "NORMAL";
         if (totalHours > 40) status = "SOBRECARGA";
@@ -81,25 +78,23 @@ public class TeacherDashboardService {
                 .teacherId(teacher.getId())
                 .fullName(teacher.getFirstName() + " " + teacher.getLastName())
                 .email(teacher.getEmail())
-                .subjects(teacher.getSubjects()) // Las materias que sabe
+                .subjects(teacher.getSubjects())
                 .activeCourses(courseInfos)
-                .totalWeeklyHours(totalHours)
+                .totalWeeklyHours(totalHours) // El DTO debe aceptar Double
                 .workloadStatus(status)
                 .build();
     }
 
+    // CAMBIO 4: El parámetro maxHours lo dejamos en int o double según prefieras, pero la lógica interna es double
     public List<AvailableTeacherDTO> findAvailableTeachers(Subject subject, int maxHours) {
-        // 1. Traer todos los profes
         List<User> allTeachers = userRepository.findByRole(Role.PROFESSOR);
         List<AvailableTeacherDTO> result = new ArrayList<>();
 
         for (User teacher : allTeachers) {
-            // Filtro 1: ¿Sabe la materia?
             if (teacher.getSubjects() == null || !teacher.getSubjects().contains(subject)) {
                 continue;
             }
 
-            // Filtro 2: Calcular carga actual (Reusamos lógica interna o simplificamos)
             List<CourseParticipation> active = participationRepository
                     .findAllByUserIdAndRolesInAndStatus(
                             teacher.getId(),
@@ -107,35 +102,32 @@ public class TeacherDashboardService {
                             ParticipationStatus.ACTIVE
                     );
 
-            long hours = 0;
+            // CAMBIO 5: Cálculo con decimales
+            double hours = 0.0;
             for(CourseParticipation p : active) {
                 if(p.getAcademicPeriod().getSchedule() != null) {
                     hours += p.getAcademicPeriod().getSchedule().getDurationInHours();
                 }
             }
 
-            // Filtro 3: ¿Tiene espacio?
             if (hours < maxHours) {
                 result.add(AvailableTeacherDTO.builder()
                         .id(teacher.getId())
                         .fullName(teacher.getFirstName() + " " + teacher.getLastName())
-                        .currentHours(hours)
+                        .currentHours(hours) // El DTO debe aceptar Double
                         .status(hours > 35 ? "ALERTA" : "LIBRE")
                         .build());
             }
         }
         return result;
     }
-    // ... dentro de TeacherDashboardService
 
     public TeacherDashboardDTO getTeacherStats(Long teacherId) {
         User teacher = userRepository.findById(teacherId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Profesor no encontrado"));
 
-        // Opcional: Validar que sea realmente un profesor
         if (teacher.getRole() != Role.PROFESSOR) {
-            // Puedes decidir si lanzar error o devolverlo igual
-            // throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "El usuario no es profesor");
+            // Validación opcional
         }
 
         return buildTeacherStats(teacher);
